Share SpellFlags
