+++ SharedSE/SharedSE.vcxitems.filters
@@ -354,6 +354,9 @@
     <ClInclude Include="$(MSBuildThisFileDirectory)TES3WeaponFlags.h">
       <Filter>Header Files\TES3</Filter>
     </ClInclude>
+    <ClInclude Include="$(MSBuildThisFileDirectory)TES3SpellFlags.h">
+      <Filter>Header Files\TES3</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="$(MSBuildThisFileDirectory)NIMatrix33.cpp">
